Update x86 crate to latest version.

+++ kernel/src/arch/x86_64/mod.rs
@@ -259,7 +259,7 @@ fn _start(argc: isize, _argv: *const *const u8) -> isize {
 
     klogger::init(
         crate::CMDLINE.get().map(|c| c.log_filter).unwrap_or("info"),
-        debug::SERIAL_PRINT_PORT.load(Ordering::Relaxed),
+        debug::SERIAL_PRINT_PORT.load(Ordering::Relaxed) as u64,
     )
     .expect("Can't set-up logging");
